requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
ci-info==0.3.0
click==8.1.6
colorama==0.4.6
configobj==5.0.8
configparser==6.0.0
etelemetry==0.3.0
ffmpeg-python==0.2.0
filelock==3.12.2
future==0.18.3
httplib2==0.22.0
idna==3.4
isodate==0.6.1
looseversion==1.3.0
lxml==4.9.3
networkx==3.1
nibabel==5.1.0
nipype==1.8.6
numpy==1.25.1
packaging==23.1
pandas==2.0.3
pathlib==1.0.1
Pillow==10.0.0
prov==2.0.0
pydot==1.4.2
pyee==11.0.0
PyMuPDF==1.22.5
pyparsing==3.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
pytz==2023.3
pyxnat==1.6
rdflib==6.3.2
requests==2.31.0
scipy==1.11.1
simplejson==3.19.1
six==1.16.0
tqdm==4.65.0
traits==6.3.2
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
you-get==0.4.1650
youtube-dl==2021.12.17
